# perf: add metrics to CopyExec and ScanExec