--- a/native/core/src/execution/jni_api.rs
+++ b/native/core/src/execution/jni_api.rs
@@ -363,7 +363,7 @@ pub unsafe extern "system" fn Java_org_apache_comet_Native_executePlan(
                     if exec_context.debug_native {
                         if let Some(plan) = &exec_context.root_op {
                             let formatted_plan_str =
-                                DisplayableExecutionPlan::with_full_metrics(plan.as_ref())
+                                DisplayableExecutionPlan::with_metrics(plan.as_ref())
                                     .indent(true);
                             info!("Comet native query plan with metrics:\n {formatted_plan_str:}");
                         }